fixed warning treated as error

+++ windows/command.cpp
@@ -4,5 +4,5 @@ Command::Command(std::vector<uint8_t>& value, int new_result_length)
 {
     byte_list = value;
     result_length = new_result_length;
-    command_length = byte_list.size();
+    command_length = (int)byte_list.size();
 }
